Fix jnaerator-runtime related build issue.
It was tricky. jnaerator-runtime, which is required to use with
JNAerator Gradle plugin, requires jna runtime, which is based on JAR.
On the other hand, in fluidsynth-kt, we reference jna runtime in AAR.
Then they caused conflict at app build time.

To resolve the puzzle, now we take these steps:

- compile fluidsynth-jna with `compileOnly`
- compile fluidsynth-kt with `implementation jnaerator-runtime` with
  `exclude (jna)`

Ideally fluidsynth-kt should probably form a Kotlin MPP and diverge
Desktop and Android implementations, but so far `app` finally builds.